Add InfraDiff and InfraObjectDiff classes
